Clarify that a project with webpack is needed
When first reading this README i wondered if it would work for my project (pure Node, no webpack). I though it would maybe be able to use webpack under the hoods to do the calculation but without requiring it in my own project. Then I tested it, and didn't work. So assuming our own project MUST indeed have webpack already and making it clear on the Readme so future readers don't wonder and test the same.